requirements.txt
cma==4.0.0
einops==0.8.1
huggingface_hub==0.30.1
matplotlib==3.6.3
numpy==1.22.4
opencv_python==4.10.0.84
Pillow==12.0.0
pydub==0.25.1
scikit_image==0.19.0
scikit_learn==1.0.1
scipy==1.7.1
soundfile==0.12.1
sympy==1.13.3
timm==0.6.5
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.66.5
Wand==0.6.13
